+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Add `get_bucket_name_or_none_from_bucket_id` to `AccountInfo` and `Cache`
 
+### Fixed
+* Use proper error handling in `ScanPoliciesManager`
+
 ## [1.7.0] - 2021-04-22
 
 ### Added